deploy tweaks

+++ db/migrate/20080910160710_wipe_all_data.rb
@@ -2,10 +2,10 @@ class WipeAllData < ActiveRecord::Migration
   def self.up
     Post.destroy_all
     Event.destroy_all
-    Topic.destroy_all
-    PostAddedToTopicEvent.destroy_all
-    PostCategorizationEvent.destroy_all
-    PostRemovedFromTopicEvent.destroy_all
+    Topic.destroy_all if Topic
+    PostAddedToTopicEvent.destroy_all  if PostAddedToTopicEvent
+    PostCategorizationEvent.destroy_all if PostCategorizationEvent
+    PostRemovedFromTopicEvent.destroy_all if PostRemovedFromTopicEvent
   end
 
   def self.down
